Adding some comments and efficiency tweaks

+++ src/minecraft/net/machinemuse/general/gui/MuseGui.java
@@ -34,30 +34,31 @@ public class MuseGui extends GuiScreen {
 	protected long creationTime;
 	protected int xSize;
 	protected int ySize;
-	
+
 	protected List<IGuiFrame> frames;
-	
+
 	public MuseGui() {
 		super();
 		frames = new ArrayList();
 	}
-	
+
 	/**
 	 * Adds the buttons (and other controls) to the screen in question.
 	 */
-	@Override public void initGui() {
+	@Override
+	public void initGui() {
 		super.initGui();
 		this.controlList.clear();
 		Keyboard.enableRepeatEvents(true);
 		creationTime = System.currentTimeMillis();
-		
+
 		int xpadding = (width - getxSize()) / 2;
 		int ypadding = (height - ySize) / 2;
 		ul = new FlyFromPointToPoint2D(absX(0), absY(0), absX(-1), absY(-1),
 				150);
 		br = new FlyFromPointToPoint2D(absX(0), absY(0), absX(1), absY(1), 150);
 	}
-	
+
 	/**
 	 * Draws the gradient-rectangle background you see in the TinkerTable gui.
 	 */
@@ -69,6 +70,7 @@ public class MuseGui extends GuiScreen {
 				new Colour(0.0F, 0.2F, 0.0F, 0.8F),
 				this.zLevel, 8f);
 	}
+
 	/**
 	 * Adds a frame to this gui's draw list.
 	 * 
@@ -77,7 +79,7 @@ public class MuseGui extends GuiScreen {
 	public void addFrame(IGuiFrame frame) {
 		frames.add(frame);
 	}
-	
+
 	/**
 	 * Draws all clickables in a list!
 	 */
@@ -93,7 +95,7 @@ public class MuseGui extends GuiScreen {
 			clickie.draw();
 		}
 	}
-	
+
 	/**
 	 * Draws the background layer for the GUI.
 	 */
@@ -101,11 +103,12 @@ public class MuseGui extends GuiScreen {
 		this.drawDefaultBackground(); // Shading on the world view
 		this.drawRectangularBackground(); // The window rectangle
 	}
-	
+
 	/**
 	 * Called every frame, draws the screen!
 	 */
-	@Override public void drawScreen(int x, int y, float z) {
+	@Override
+	public void drawScreen(int x, int y, float z) {
 		super.drawScreen(x, y, z);
 		update();
 		drawBackground();
@@ -114,6 +117,7 @@ public class MuseGui extends GuiScreen {
 		}
 		drawToolTip();
 	}
+
 	public void update() {
 		double x = Mouse.getEventX() * this.width / (double) this.mc.displayWidth;
 		double y = this.height - Mouse.getEventY() * this.height
@@ -122,7 +126,7 @@ public class MuseGui extends GuiScreen {
 			frame.update(x, y);
 		}
 	}
-	
+
 	/**
 	 * Returns the first ID in the list that is hit by a click
 	 * 
@@ -144,15 +148,16 @@ public class MuseGui extends GuiScreen {
 		}
 		return -1;
 	}
-	
+
 	/**
 	 * Whether or not this gui pauses the game in single player.
 	 */
-	@Override public boolean doesGuiPauseGame()
+	@Override
+	public boolean doesGuiPauseGame()
 	{
 		return false;
 	}
-	
+
 	/**
 	 * Returns absolute screen coordinates (int 0 to width) from a relative
 	 * coordinate (float -1.0F to +1.0F)
@@ -161,23 +166,23 @@ public class MuseGui extends GuiScreen {
 	 *            Relative X coordinate
 	 * @return Absolute X coordinate
 	 */
-	public int absX(float relx) {
+	public int absX(double relx) {
 		int absx = (int) ((relx + 1) * getxSize() / 2);
 		int xpadding = (width - getxSize()) / 2;
 		return absx + xpadding;
 	}
-	
+
 	/**
 	 * Returns relative coordinate (float -1.0F to +1.0F) from absolute
 	 * coordinates (int 0 to width)
 	 * 
 	 */
-	public int relX(float absx) {
+	public int relX(double absx) {
 		int padding = (width - getxSize()) / 2;
 		int relx = (int) ((absx - padding) * 2 / getxSize() - 1);
 		return relx;
 	}
-	
+
 	/**
 	 * Returns absolute screen coordinates (int 0 to width) from a relative
 	 * coordinate (float -1.0F to +1.0F)
@@ -186,12 +191,12 @@ public class MuseGui extends GuiScreen {
 	 *            Relative Y coordinate
 	 * @return Absolute Y coordinate
 	 */
-	public int absY(float rely) {
+	public int absY(double rely) {
 		int absy = (int) ((rely + 1) * ySize / 2);
 		int ypadding = (height - ySize) / 2;
 		return absy + ypadding;
 	}
-	
+
 	/**
 	 * Returns relative coordinate (float -1.0F to +1.0F) from absolute
 	 * coordinates (int 0 to width)
@@ -202,14 +207,14 @@ public class MuseGui extends GuiScreen {
 		int rely = (int) ((absy - padding) * 2 / ySize - 1);
 		return rely;
 	}
-	
+
 	/**
 	 * @return the xSize
 	 */
 	public int getxSize() {
 		return xSize;
 	}
-	
+
 	/**
 	 * @param xSize
 	 *            the xSize to set
@@ -217,14 +222,14 @@ public class MuseGui extends GuiScreen {
 	public void setxSize(int xSize) {
 		this.xSize = xSize;
 	}
-	
+
 	/**
 	 * @return the ySize
 	 */
 	public int getySize() {
 		return ySize;
 	}
-	
+
 	/**
 	 * @param ySize
 	 *            the ySize to set
@@ -232,29 +237,31 @@ public class MuseGui extends GuiScreen {
 	public void setySize(int ySize) {
 		this.ySize = ySize;
 	}
-	
+
 	/**
 	 * Called when the mouse is clicked.
 	 */
-	@Override protected void mouseClicked(int x, int y, int button)
+	@Override
+	protected void mouseClicked(int x, int y, int button)
 	{
 		for (IGuiFrame frame : frames) {
 			frame.onMouseDown(x, y, button);
 		}
 	}
-	
+
 	/**
 	 * Called when the mouse is moved or a mouse button is released. Signature:
 	 * (mouseX, mouseY, which) which==-1 is mouseMove, which==0 or which==1 is
 	 * mouseUp
 	 */
-	@Override protected void mouseMovedOrUp(int x, int y, int which)
+	@Override
+	protected void mouseMovedOrUp(int x, int y, int which)
 	{
 		for (IGuiFrame frame : frames) {
 			frame.onMouseUp(x, y, which);
 		}
 	}
-	
+
 	/**
 	 * @param x
 	 * @param y
@@ -282,11 +289,11 @@ public class MuseGui extends GuiScreen {
 			} else {
 				top = y;
 				bottom = y + 10 * tooltip.size() + 8;
-				
+
 				left = x + 4;
 				right = x + 12 + strwidth;
 			}
-			
+
 			MuseRenderer.drawFrameRect(
 					left, top,
 					right, bottom,
@@ -300,7 +307,7 @@ public class MuseGui extends GuiScreen {
 			}
 		}
 	}
-	
+
 	/**
 	 * @return
 	 */
@@ -314,9 +321,10 @@ public class MuseGui extends GuiScreen {
 		}
 		return null;
 	}
-	
+
 	/**
 	 * 
 	 */
-	public void refresh() {}
+	public void refresh() {
+	}
 }
